linear-search-in-C
Explanation:

linearSearch function: This function takes an array arr of integers, its size n, and the value to be searched x as inputs. It uses the linear search algorithm to find the index of x in arr. The algorithm simply iterates over the elements of arr one by one, checking if each element is equal to x. The function returns the index of x if found, and -1 otherwise.

main function: This function creates an array of integers arr, initializes it with some values, and calls the linearSearch function to search for the value x. If the function returns -1, it means the value was not found in the array, and the program prints "Element not found". Otherwise, it prints "Element found at index index"