Create jump-search
jumpSearch function: This function takes a sorted array arr of integers, its size n, and the value to be searched x as inputs. It uses the jump search algorithm to find the index of x in arr. The algorithm works by first dividing the array into blocks of size sqrt(n), and then jumping to the beginning of the next block until an element greater than or equal to x is found or the end of the array is reached. Once a block is found that contains x or may contain x, a linear search is performed within that block to find the index of x. The function returns the index of x if found, and -1 otherwise.

main function: This function creates a sorted array of integers arr, initializes it with some values, and calls the jumpSearch function to search for the value x. If the function returns -1, it means the value was not found in the array, and the program prints "Element not found". Otherwise, it prints "Element found at index index".